Publish to <https://nextstrain.github.io/status/> every 5 minutes
Currently the job takes ~3 min to run, so it should be ok to schedule it
to run every 5.  The concurrency group will ensure it doesn't stomp on
itself, but it also means if the job starts taking longer than 5 min to
run then jobs will start queueing up and becoming further out of date.
We should endeavor to make the job as quick as possible.  3 min already
seems on the slow end! but there's a lot of (paginated) API requests
happening behind the scenes, some of which require client-side
filtering.

This workflow will also function as CI for in-development branches and
PRs.  Those runs won't update the website, but build artifacts are still
uploaded to the workflow run for manual inspection.  Including
github.ref in the concurrency group ensures that non-publishing jobs
don't queue up behind publishing ones.