Query Steampipe thru psql for more flexibility
The `steampipe query` interface was alright, but the psql interface
handles multi-statement input and thus lets us do more, like set Pg
output/formatting options such as intervalstyle.  psql is also a very
well-worn and time-tested tool and so has greater and better resources
available for how to do things with it, both within our team and on the
internet at large.