Wait for Steampipe plugin schemas to become available before running psql
Steampipe is sometimes slow to register schemas for plugins at Postgres
server start.  Or maybe Postgres is slow to import the foreign table
schemas.  Either way, while locally it was always fast enough, in CI it
was often losing the race with our query and causing errors.

Ensure the Steampipe plugin connections (and thus, it seems, schemas)
are in the "ready" state.  This does require that any plugins which are
installed must be set up correctly and able to reach the "ready" state.