Inhibit transitions during initial page load
If we're loading into the time-relative layout, we don't want the
transition but instead to look as if we loaded into that layout
directly.

I had the status page visible in the background on the time-relative
layout (most useful to me recently) and the flash of transitions on the
auto-refresh was bothering me.