CI: Never regenerate favicon-*.svg from source
While it may be useful to regenerate them in development—it certainly
was for me!—doing so in CI doesn't work because Inkscape isn't
installed.  We could install Inkscape, but why bother?  The generated
copies that are checked in to source control can be used as-is.

Suggested by @joverlee521 during troubleshooting¹ why CI sometimes
failed: it was sometimes trying to regenerate these files because
sometimes (it seems?) the input mtimes were newer than the output
mtimes.

¹ <https://github.com/nextstrain/status/pull/29#issuecomment-2474677671>