Indicate non-underlined links with underlines on hover
The underline is such a universal hint that a thing is clickable.  It
also clearly demarcates the extent of the link, which is helpful for
adjacent or nearly adjacent links that go to different places.

While the underline feels too noisy to keep always present for these
links, showing it on hover seems like a good compromise.