Indicate in nav if any of a pathogen's most recent workflow runs have failed
Helps quickly jump to the failing pathogen.  For example, if you notice
the favicon indicates failure and want to go see what's up, you don't
have to scroll thru the page to find the failure(s).

Other conclusion/status indicators are not shown in nav because it
seemed too noisy and not notable enough.