Gracefully handle being offline
Inhibit the automatic refresh when offline to avoid triggering a "this
site cannot be loaded" error page in the browser.  That's game over for
us: once it happens we're unloaded and cannot recover with a subsequent
refresh.

Display a prominent message when offline to avoid misleading a viewer
who may not realize they're offline and refreshes aren't happening.
This covers about the only useful case of the browser error page.

Firefox has a "work offline" mode which is useful for testing this, but
the setting can also be flipped other ways (and in other browsers).