Use a customized github plugin for Steampipe
Includes the following changes:

  - <https://github.com/turbot/steampipe-plugin-github/pull/463>
  - <https://github.com/turbot/steampipe-plugin-github/pull/464>
  - <https://github.com/turbot/steampipe-plugin-github/pull/465>

for additional functionality around run attempts (to be used in the next
commit) and improved efficiency listing runs (automatically!).

The hope is that I can get my changes upstreamed and we can stop using a
customized and vendored version, but I'm not sure if that's likely or
not.