Improve handling of workflow runs with multiple attempts (i.e. re-runs)
The query now distinguishes between the "overall" run record—which is
an aggregate over the first and last run attempts—and the individual run
attempt records.

For our purposes, it's sensible to display re-runs as totally
independent runs and so that's what we do.  This increases the
information on the status page and better reflects reality (e.g. re-runs
don't mask a previous failure).  It's also the simplest treatment of
re-runs that doesn't pretend they don't exist.

The "run details" link out to GitHub now always points to the specific
attempt's page.

Requires (for now) our customized version of the github plugin for
Steampipe that's set up in the previous commit.

Resolves <https://github.com/nextstrain/status/issues/18>.