steampipe-plugins/github: Check HTTP status code, not a stringy error message
This checked started failing because (it seems) GitHub changed their
error message string from "No commit found" to "No object found".

Checking stringy error messages is doomed to this sort of problem from
the start.  This, unfortunately, is just one place that's done: there
are many others in this codebase (grep for "err.Error").  I'm leaving
them be for now, but they suffer from the same bug waiting to happen.

Resolves <https://github.com/nextstrain/status/issues/36>.