Warn about page staleness, e.g. due to updates failing to happen
No updates in this case might be that periodic refreshing is paused
(e.g. because you have a <details> element open and maybe you don't
realize it) or that periodic regeneration of the page is not happening
(e.g. scheduled jobs are failing).  Either way, it's good to call out
that the time of generation is now older than expected.  30 minutes
seems like a reasonable threshold for now given the usual update cadence
of this page and the workflows it's showing.

Combines the staleness warning with the offline warning, if either is
visible, so that there's just one (orange)red warning box at the top of
the page instead of two or something.