Correctly sort runs by created_at (descending) as intended
created_at is an ISO 8601 timestamp, not an integer (i.e. Unix epoch
time), so applying the unary negation operator to it resulted in a bunch
of NaNs.  These were ignored by sort() and so the reverse chronological
order we saw in practice was based solely on run ids.  Whoops!  Convert
created_at to an epoch time for sorting.

I noticed this due to a layout bug where two runs occupied the same
space in the time-relative layout.  That was because the day transition
detection in that layout mode assumes (requires) a correct ordering by
created_at when the ids are not sufficient to do so (e.g. because of a
re-run attempt a few days later when another separate run also exists on
that day).  I knew of this assumption/requirement but figured it
warranted explicitly commenting on too.